+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   communicate that they come with padding. `minor-modes` sets this to
   `left`. This value is respected and no additional padding is added
   to segments to the left or right of it.
+- Space calculations for windows are now cached, greatly reducing the
+  time spent on figuring out which segments can be shown.
 
 ### Changed